remove components.js and DisplayFourBooks.jsx, mobile optimize landing, separate App.css into layout and global styles sections
