+++ src/components/components.js
@@ -1,8 +1,7 @@
 import Book from './Book.jsx';
-import DisplayFourBooks from './DisplayFourBooks.jsx';
 import Footer from './Footer';
 import Nav from './Nav.jsx';
 import Price from './Price';
 import Rating from './Rating.jsx';
 
-export { Nav, Rating, Book, Price, DisplayFourBooks, Footer };
+export { Nav, Rating, Book, Price, Footer };
