+++ src/pages/BookInfo/BookInfo.jsx
@@ -4,7 +4,6 @@ import { useParams } from 'react-router';
 import { Link } from 'react-router-dom';
 
 // components
-import { DisplayFourBooks } from 'components/components';
 import BookSelected from './BookSelected';
 
 // styles & assets
@@ -13,7 +12,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import './BookInfo.css';
 
 export default function BookInfo({ addItemToCart }) {
-  const { id } = useParams();
+  const { id: currentBookID } = useParams();
   const { docs: books } = useCollection('books');
   const [book, setBook] = useState(null);
 
@@ -40,7 +39,13 @@ export default function BookInfo({ addItemToCart }) {
           </div>
           <div className='row recc-books'>
             <h2 className='recc-books__title'>Recommended Books</h2>
-            {books && <DisplayFourBooks id={id} />}
+            <div className='books__container'>
+              {books &&
+                books
+                  .filter((book) => book.rating === 5 && book.id !== currentBookID) // if they are on book info page then show only new books
+                  .slice(0, 4)
+                  .map((book) => <Book book={book} key={book.id} />)}
+            </div>
           </div>
         </div>
       </main>
