+++ src/pages/Home/Landing.css
@@ -8,13 +8,13 @@ header .container {
 }
 
 .header__title {
-  font-size: 3.5rem;
+  font-size: 56px;
   color: #7342D6;
   margin-bottom: 24px;
 }
 
 .header__sub-title {
-  font-size: 1.5rem;
+  font-size: 24px;
   margin-bottom: 32px;
 }
 
@@ -35,5 +35,15 @@ header .container {
 @media screen and (max-width: 576px) {
   .header__img {
     height: 35vh;
+  }
+}
+
+@media screen and (max-width: 480px) {
+  .header__title {
+    font-size: 36px;
+  }
+
+  .header__sub-title {
+    font-size: 20px;
   }
 }
